backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
asyncpg==0.29.0
boto3==1.35.36
sse-starlette==2.1.3
python-dotenv==1.0.1
CouchDB3>=2.0.0

